// networking/sylius-api-client/.ts.schemas.ts
export type RankingProductVintageRatingReadModelCollection200HydraView = {
  "@id"?: string;
  "@type"?: string;
  "hydra:first"?: string;
  "hydra:last"?: string;
  "hydra:next"?: string;
  "hydra:previous"?: string;
};

export type RankingProductVintageRatingReadModelCollection200HydraSearchHydraMappingItem = {
  "@type"?: string;
  property?: string | null;
  required?: boolean;
  variable?: string;
};

export type RankingProductVintageRatingReadModelCollection200HydraSearch = {
  "@type"?: string;
  "hydra:mapping"?: RankingProductVintageRatingReadModelCollection200HydraSearchHydraMappingItem[];
  "hydra:template"?: string;
  "hydra:variableRepresentation"?: string;
};

export type RankingProductVintageRatingReadModelCollection200 = {
  "hydra:member": ProductVintageRatingReadModelJsonldShopProductVintageRatingRankingRead[];
  "hydra:search"?: RankingProductVintageRatingReadModelCollection200HydraSearch;
  "hydra:totalItems"?: number;
  "hydra:view"?: RankingProductVintageRatingReadModelCollection200HydraView;
};
// ...
export type GetAuctionItemReadModelCollection200HydraView = {
  "@id"?: string;
  "@type"?: string;
  "hydra:first"?: string;
  "hydra:last"?: string;
  "hydra:next"?: string;
  "hydra:previous"?: string;
};

export type GetAuctionItemReadModelCollection200HydraSearchHydraMappingItem = {
  "@type"?: string;
  property?: string | null;
  required?: boolean;
  variable?: string;
};

export type GetAuctionItemReadModelCollection200HydraSearch = {
  "@type"?: string;
  "hydra:mapping"?: GetAuctionItemReadModelCollection200HydraSearchHydraMappingItem[];
  "hydra:template"?: string;
  "hydra:variableRepresentation"?: string;
};

export type GetAuctionItemReadModelCollection200 = {
  "hydra:member": AuctionItemReadModelJsonldShopAuctionItemReadModelTopAdjudicationsRead[];
  "hydra:search"?: GetAuctionItemReadModelCollection200HydraSearch;
  "hydra:totalItems"?: number;
  "hydra:view"?: GetAuctionItemReadModelCollection200HydraView;
};
// ...
export interface ProductVintageRatingReadModelJsonldShopProductVintageRatingRankingRead {
  readonly "@id"?: string;
  readonly "@type"?: string;
  productVintage?: ProductVintageJsonldShopProductVintageRatingRankingRead;
  value?: number;
}
// ...
export interface ProductVintageRatingReadModelJsonldShopProductVintageRatingForEstateRead {
  readonly "@id"?: string;
  readonly "@type"?: string;
  readonly id?: number;
  productVintage?: ProductVintageJsonldShopProductVintageRatingForEstateRead;
  value?: number;
}
// ...
export interface ProductVintageJsonldShopProductVintageRatingRankingRead {
  readonly "@context"?: ProductVintageJsonldShopProductVintageRatingRankingReadContext;
  readonly "@id"?: string;
  readonly "@type"?: string;
  code?: string;
  product?: ProductJsonldShopProductVintageRatingRankingRead;
  year?: number | null;
}
// ...
export interface ProductVintageJsonldShopProductVintageRatingForEstateRead {
  readonly "@context"?: ProductVintageJsonldShopProductVintageRatingForEstateReadContext;
  readonly "@id"?: string;
  readonly "@type"?: string;
  code?: string;
  readonly id?: number;
  product?: ProductJsonldShopProductVintageRatingForEstateRead;
  year?: number | null;
}
// ...
export type ProductVintageJsonldShopProductVintageRatingForEstateReadContext =
  | string
  | ProductVintageJsonldShopProductVintageRatingForEstateReadContextOneOf;
// ...
export type ProductVariantInAuctionCatalogJsonldShopAuctionItemReadModelTopAdjudicationsReadContext =

    | string
    | ProductVariantInAuctionCatalogJsonldShopAuctionItemReadModelTopAdjudicationsReadContextOneOf;

export interface ProductVariantInAuctionCatalogJsonldShopAuctionItemReadModelTopAdjudicationsRead {
  readonly "@context"?: ProductVariantInAuctionCatalogJsonldShopAuctionItemReadModelTopAdjudicationsReadContext;
  readonly "@id"?: string;
  readonly "@type"?: string;
  productVariant?: ProductVariantJsonldShopAuctionItemReadModelTopAdjudicationsRead;
}
// ...
export interface ProductVariantJsonldShopAuctionItemReadModelTopAdjudicationsRead {
  readonly "@context"?: ProductVariantJsonldShopAuctionItemReadModelTopAdjudicationsReadContext;
  readonly "@id"?: string;
  readonly "@type"?: string;
  code: string;
  format?: ProductVariantJsonldShopAuctionItemReadModelTopAdjudicationsReadFormat;
  numberOfBottles?: number;
  product?: ProductJsonldShopAuctionItemReadModelTopAdjudicationsRead;
  productVintage?: ProductVariantJsonldShopAuctionItemReadModelTopAdjudicationsReadProductVintage;
}
// ...
export interface ProductJsonldShopProductVintageRatingForEstateRead {
  readonly "@context"?: ProductJsonldShopProductVintageRatingForEstateReadContext;
  readonly "@id"?: string;
  readonly "@type"?: string;
  appellation?: string | null;
  color?: ProductJsonldShopProductVintageRatingForEstateReadColor;
  estate?: ProductJsonldShopProductVintageRatingForEstateReadEstate;
  name?: string | null;
  region?: ProductJsonldShopProductVintageRatingForEstateReadRegion;
}
// ...
export interface AuctionItemReadModelJsonldShopAuctionItemReadModelTopAdjudicationsRead {
  readonly "@id"?: string;
  readonly "@type"?: string;
  endDate?: string | null;
  highestBid?: number | null;
  productVariantInAuctionCatalog?: ProductVariantInAuctionCatalogJsonldShopAuctionItemReadModelTopAdjudicationsRead;
}